pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "xstage"
version = "0.1.0"
description = "Professional USD viewer with pipeline integration"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "NOX VFX", email = "pipeline@nox-vfx.com"}
]
keywords = ["usd", "openusd", "vfx", "pipeline", "viewer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: POSIX :: Linux",
    "Operating System :: POSIX :: Linux :: Ubuntu",
    "Operating System :: POSIX :: Linux :: Red Hat",
]
dependencies = [
    "usd-core>=23.11",
    "PySide6>=6.6.0",
    "PySide6-Addons>=6.6.0",
    "PyOpenGL>=3.1.7",
    "PyOpenGL-accelerate>=3.1.7",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "psutil>=5.9.0",
    "pygltflib>=1.16.0",
    "trimesh>=3.15.0",
    "alembic>=1.8.0; python_version >= '3.9'",
]

[project.urls]
Homepage = "https://github.com/xstage-pipeline/xstage"
Documentation = "https://xstage-pipeline.github.io"
Repository = "https://github.com/xstage-pipeline/xstage"
Issues = "https://github.com/xstage-pipeline/xstage/issues"

[project.scripts]
xstage = "xstage.core.viewer:main"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pylint]
max-line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
timeout = 300
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gui: marks tests that require GUI (deselect with '-m \"not gui\"')",
    "integration: marks tests as integration tests",
]